optimizing type collections

+++ pyrestsdk/request/_base_request310.py
@@ -2,39 +2,24 @@
 
 from __future__ import annotations
 
-from typing import (
-    TypeVar,
-    Union,
-    Optional,
-    Iterable,
-    Dict,
-    Any,
-)
-
+from typing import TypeVar, Union, Optional, Iterable, Dict, Any
 import logging
 
 from requests import Response
 
 from pyrestsdk.type.enum import HttpsMethod
-from pyrestsdk.type.model import (
-    Entity,
-    QueryOption,
-    HeaderOption,
-)
-
+from pyrestsdk.type.model import Entity, QueryOption, HeaderOption
 from pyrestsdk.request._request import Request
 
 Logger = logging.getLogger(__name__)
 
 T = TypeVar("T", bound=Entity)
-Q = TypeVar("Q", bound=QueryOption)
-H = TypeVar("H", bound=HeaderOption)
 
 
 class BaseRequest(Request[T]):
     """The Base Request Type"""
 
-    def _parse_options(self, options: Optional[Iterable[Union[Q, H]]]) -> None:
+    def _parse_options(self, options: Optional[Iterable[Union[QueryOption, HeaderOption]]]) -> None:
         """Parses the provided options into either header or query options"""
 
         Logger.info("%s._parse_options: function called", type(self).__name__)
@@ -43,20 +28,16 @@ class BaseRequest(Request[T]):
             return None
 
         for option in options:
-            match type(option):
-                case n if issubclass(n, HeaderOption):
-                    self.header_options.append(option)
-                case n if issubclass(n, QueryOption):
-                    self.query_options.append(option)
-                case other:
-                    raise TypeError(
-                        "Unexpected type: %s, expected subtype of HeaderOption or QueryOption",
-                        type(other),
-                    )
+            if isinstance(option, HeaderOption):
+                self.header_options.append(option)
+            elif isinstance(option, QueryOption):
+                self.query_options.append(option)
+            else:
+                raise TypeError(
+                    f"Unexpected type: {type(option)}, expected instance of HeaderOption or QueryOption"
+                )
 
-    def _send_request(
-        self, args: Dict[str, Any], value: Optional[Union[T, Dict[str, Any], str]]
-    ) -> Optional[Response]:
+    def _send_request(self, args: Dict[str, Any], value: Optional[Union[T, Dict[str, Any], str]]) -> Optional[Response]:
         """Sends request
 
         Args:
@@ -76,15 +57,14 @@ class BaseRequest(Request[T]):
             self.request_method.name,
         )
 
-        match self.request_method:
-            case HttpsMethod.GET:
-                return self._client.get(**args)
-            case HttpsMethod.POST:
-                return self._client.post(**args)
-            case HttpsMethod.DELETE:
-                self._client.delete(**args)
-                return None
-            case HttpsMethod.PUT:
-                return self._client.put(**args)
-            case other:
-                raise TypeError(f"Unknown HTTPS method {other.name}")
+        if self.request_method == HttpsMethod.GET:
+            return self._client.get(**args)
+        elif self.request_method == HttpsMethod.POST:
+            return self._client.post(**args)
+        elif self.request_method == HttpsMethod.DELETE:
+            self._client.delete(**args)
+            return None
+        elif self.request_method == HttpsMethod.PUT:
+            return self._client.put(**args)
+        else:
+            raise TypeError(f"Unknown HTTPS method {self.request_method.name}")
